dc_outbound/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
simplejson
pandas
regex
protobuf==3.20.0